+++ src/main/java/woowacourse/shoppingcart/exception/CustomerInfoNotMatchException.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.exception;
+package woowacourse.shoppingcart.exception;
 
 public class CustomerInfoNotMatchException extends RuntimeException {
