+++ src/test/java/woowacourse/shoppingcart/ui/CustomerControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.jdbc.Sql;
-import woowacourse.auth.exception.PasswordNotMatchException;
+import woowacourse.shoppingcart.exception.PasswordNotMatchException;
 import woowacourse.shoppingcart.dto.CustomerDeleteRequest;
 import woowacourse.shoppingcart.dto.CustomerDetailResponse;
 import woowacourse.shoppingcart.dto.CustomerPasswordUpdateRequest;
